chore(node): cleanup node cli
