+++ bin/node/src/flags/mod.rs
@@ -1,0 +1,7 @@
+//! CLI Flags
+
+mod globals;
+pub use globals::GlobalArgs;
+
+mod p2p;
+pub use p2p::P2PArgs;
